feat: Add UserList component
This commit adds the UserList component to the project. The UserList component is responsible for rendering a table of users, including their names and email addresses. It receives an array of users as a prop and dynamically generates the table rows based on the data. This component will be used in the App component to display a list of users.